add note that fm only works in zsh
the fm script only works in zsh, since it does not display lines
properly in bash.  the '-k' option is also not available in the built-in
'read' command in bash.

adding the '-1' option to the 'ls' commands did not fix the first
problem.

using the '-n' option with 'read' works in bash, but the shell must be
detected first, since this is not available in zsh.  detecting the shell
pid can be done using the command:

  ps -o args= -p "$$"