feature:  add search functionality
the '/' and '?' key bindings have been added, which give the user the
ability to search for a pattern.

the 'n' and 'N' key bindings have been added, which go to the next and
previous pattern respectively.

the find functions (for the 'f' and 'F' key bindings) have also been
changed so that all searching works in a similar way.