simplify the code
the '$previous_directory' variable is assigned before the 'read_keys'
function is called, so assigning a value to it if it's not already
assigned was not necessary.

also, if the variable doesn't have a value, the 'go_to' function will
not move the current directory to any other directory, so this wasn't
needed even if the variable was not already assigned.