simplify the code
move some lines from the 'shell_command' function and from the lines for
redrawing the screen to a new 'stay_on_same_line' function since the
functionality was the same.

redrawing the screen now also tries to first move the line to the same
file name instead of trying to move it to the same line number.